chore: refactor toggle-timeframe-component
