+++ src/molecules/chart/Chart.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 
 import Box from "@mui/material/Box"
 
@@ -13,10 +13,6 @@ import {
 	YAxis
 } from "recharts";
 
-// MUI Imports
-import ToggleButton from "@mui/material/ToggleButton";
-import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
-
 import moment from "moment";
 
 // Custom Imports
@@ -25,101 +21,17 @@ import CustomChartTooltip from "/src/atoms/custom-chart-tooltip/CustomChartToolt
 
 import CompleteLogoChart from "/src/ions/img/chart/complete-logo-chart.svg"
 
-import useStore from "/src/ions/hooks/state/useStore"
+import ToggleTimeframeComponent from "/src/molecules/toggle-timeframe-component/ToggleTimeframeComponent"
+
 
 const Chart = ({ dataArray }) => {
 
-	const loading = useStore((state) => state.loading)
-
-	const setTimeFrame = useStore((state) => state.setTimeFrame);
-	const setInterval = useStore((state) => state.setInterval);
-
-	const [alignment, setAlignment] = useState(1);
-
-	const handleChange = (event, newAlignment) => {
-		setAlignment(newAlignment)
-	}
-
-	useEffect(()=>{
-		setAlignment(7);
-		setInterval("daily");
-		setTimeFrame(7)
-	},[setInterval, setTimeFrame])
-
 	return (
 		<Box sx={{ width: "100%", display: "flex", flexDirection: "column", p: .5}}>
-			{/*Time-frame component*/}
-			<Box sx={{ m: .5, alignSelf: "flex-end" }}>
-				<ToggleButtonGroup
-					exclusive
-					color="primary"
-					value={alignment}
-					disabled={loading}
-					size="small"
-					onChange={handleChange}
-				>
-					<ToggleButton
-						variant="outlined"
-						value={1}
-						onClick={()=>{
-							setTimeFrame(1);
-							setInterval("hourly")
-						}}
-					>
-						24H
-					</ToggleButton>
-					<ToggleButton
-						variant="outlined"
-						value={7}
-						onClick={()=>{
-							setTimeFrame(7);
-							setInterval("daily")
-						}}
-					>
-						7D
-					</ToggleButton>
-					<ToggleButton
-						variant="outlined"
-						value={14}
-						onClick={()=>{
-							setTimeFrame(14);
-							setInterval("daily")
-						}}
-					>
-						14D
-					</ToggleButton>
-					<ToggleButton
-						variant="outlined"
-						value={30}
-						onClick={()=>{
-							setTimeFrame(30);
-							setInterval("daily")
-						}}
-					>
-						1M
-					</ToggleButton>
-					<ToggleButton
-						variant="outlined"
-						value={90}
-						onClick={()=>{
-							setTimeFrame(90);
-							setInterval("daily")
-						}}
-					>
-						3M
-					</ToggleButton>
-					<ToggleButton
-						variant="outlined"
-						value={360}
-						onClick={()=>{
-							setTimeFrame(360);
-							setInterval("monthly")
-						}}
-					>
-						1Y
-					</ToggleButton>
-				</ToggleButtonGroup>
-			</Box>
+
+			{/*Toggle-Timeframe component*/}
+			<ToggleTimeframeComponent />
+
 			<ResponsiveContainer width="100%" height={275}>
 				<AreaChart type="monotone" stroke="#8884d8" data={dataArray} margin={{ left: -60, top: 20 }}>
 					<defs>
